Backend/requirements.txt
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
Pillow==12.2.0
pytesseract==0.3.13
python-dotenv==1.2.2
requests==2.32.5
